pyproject.toml
[project]
name = "leave-policy-agent"
version = "0.1.0"
description = "Production-grade Leave Policy Assistant using Google ADK"
requires-python = ">=3.12"

dependencies = [
    # Core
    "google-adk>=1.19.0",
    "litellm>=1.80.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",

    # Snowflake
    "snowflake-snowpark-python",

    # Monitoring
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-instrumentation-fastapi",

    # Utilities
    "python-dateutil>=2.8.2",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",

    "black>=24.0.0",
    "ruff>=0.4.0",
    "isort>=5.13.0",
    "pre-commit>=3.6.0"
]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E","F","I","B","UP","ASYNC"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"

[tool.isort]
profile = "black"

[tool.ruff.lint.isort]
known-first-party = ["src"]
